removed the tri language vectors
we can do this in the main by making a vector that stores a vector of ints. argv[] - the file name, train[i] for the index of the vectors of frequency. make double sim(vector train, vector test)... test language = argc-1 (argc >=3). interpret the command line - create the vectors - compare them